refactor(Client): create event loading framework
