+++ public/structures/Client/events/error.js
@@ -1,0 +1,6 @@
+module.exports = {
+	name: 'error',
+	async execute(client, error) {
+		client.transmit(this.name, `[${error.code}] ${error.message}`)
+	},
+}
