+++ public/structures/Client/events/templateEvent
@@ -1,0 +1,7 @@
+module.exports = {
+	name: 'eventName',
+	once: true | false,
+	async execute(client, ...args) {
+		//code here
+	},
+}
